feat: change User unique index from mail to userPrincipalName
This means the lib will check for the `user.username` field to decide when to get or create a user. This is needed to work with tenant users with no email address saved.

If you want to enforce unique email addresses in your database (checking against `user.email`) please set the option `MICROSOFT_SSO_UNIQUE_EMAIL=True`